requirements.txt
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pyyaml>=6.0
transformers>=4.30.0
datasets>=4.4.1

